I have Autotools-Fu.
We should now be at the bare minimum of files under version control,
which also means that _NO_ generated files are versioned. This should
mean that any problems which arise from, for example, having different
libtool versions on my Fedora machine and my iBook, are solved.

I found out that autogen.sh is necessary after all, since Gtk-Doc and
Intltool require their own preparation programs in addition to the
Autotools' autoreconf. I rewrote it to be much simpler and easier to
read, and removed all of the checks for extreme system configurations.
That's because users shouldn't have to run it, only developers, who
should all have decent systems with decent versions of Autotools.

The autogen script runs autoreconf, but not configure. (Apparently
running configure automatically is a big timewaster.) So, after 
checking out a new revision, the correct sequence to run is:

./autogen.sh
./configure
make

There are still a couple of warnings from the Gtk-Doc autoconf macros,
but since we don't actually do anything with Gtk-Doc yet, I will save
fixing that for when I attain Autotools-Nirvana.