Sync new GStreamer element to parent state
In GStreamer 1.0, when adding a new element to the pipeline in a signal
handler, its state has to be synced to the pipeline's state, because the
handler could be called in the middle of a state change which would
cause the state change to hang.

This was supposed to be done in GStreamer 0.10 too, but it never caused
any problems.