Fix wraparound bug
Better compiler warnings turned up this bug that thankfully never
manifested itself: a calculation using unsigned ints that might have
wrapped around. It was subsequently checked if it had gone below 0, but
of course that is impossible with unsigned ints. Now we check beforehand
if the calculation would wrap around.