Don't wait for a shutdown keypress if no windows
If there were no windows open, or only graphics and blank windows, then
it could happen that when a game ended there would be no way to quit the
interpreter, because it would be waiting for a keypress that could never
come.

Instead, only wait for a keypress if at least one text grid or buffer
window is open.