Unit test runner
This glkunit-runner executable does the same thing as plugin-loader,
only without a window, and quitting immediately when the plugin stops
instead of waiting for the window to be closed.

This allows us to use glk_exit() in glkunit, instead of ANSI exit(),
which the Glk spec forbids because otherwise the Glk library can't shut
down properly.

Extract common code between glkunit-runner and plugin-loader into a
plugin-utils file.