glk: Threading without gdk_threads_enter()
gdk_threads_enter() and gdk_threads_leave() are deprecated in GTK 3.x.
The proper way to use threading with GTK programs is to avoid
manipulating the GUI from any other thread but the main thread at all,
and instead queue up GUI operations with gdk_threads_add_idle().
We could have kept our heads in the sand and not dealt with the problem
until GTK 4, except that GDK threads on OSX have not worked since GTK
3.6, and we do want this program to remain portable.

This obviously requires a massive rewrite in Chimara.

In order to avoid a big mess of idle callbacks being queued up
everywhere, this commit adds a UI message queue. The Glk thread queues up
UI messages representing common operations such as CLEAR_WINDOW,
REQUEST_LINE_INPUT, MOVE_CURSOR, etc., and stuffs them into the queue
where they are processed in order by the UI thread. The Glk thread can
choose to wait for an answer from the UI thread if the message requires
so.

ui_message_new(), ui_message_queue(), and ui_message_queue_and_await()
are meant to be called from the Glk thread. All other ui_*() functions
are meant to be called from the UI thread.

Specifically, UI messages are created in the Glk thread, then ownership
of them passes to the UI thread where they are eventually freed. No UI
message may be accessed at all by the Glk thread after it has been
queued.

The ui_*() functions are not all in the ui-*.c files yet, but that is in
order to leave as much of the code in the same place as possible so that
this diff is easier to read. I'll be moving functions around in the next
commit.