graphics: Return early if drawing 0-size image
When calling glk_draw_image_scaled() with a width or height of 0, return
early, as the spec says that nothing is drawn in that case. The spec is
not clear whether the drawing operation should be considered a success or
failure in that case, but we return FALSE and print a warning. Previously
we would also have printed warnings about trying to scale a pixbuf to 0
size, anyway.