interpreters: Update to latest versions
This updates Bocfel, Frotz, Git, Glulxe, and Nitfol to the exact versions
that Garglk bundles, as of https://github.com/garglk/garglk commit
e27d9bb906db09cb7e92ded7fcea8f04123497e6.

We remove the patches in the interpreters/ directory. Now the only
modifications are the Makefile.am files in each directory.

This update also slightly changes the behaviour of Nitfol's ignore errors
option, so change the documentation to reflect this.