player: Reorganize menus
This changes the player window to use a custom title bar. It has an "Open"
button in the top left, taking care of both the "Open" and "Open Recent"
former menu items. "Copy" and "Paste" are left out, since they work as
normal with Ctrl+C and Ctrl+V without a menu item. "Show/Hide Toolbar" is
left out as well, because there is no toolbar. The other menu items are
moved to a hamburger menu on the right hand side of the title bar.