glk: Use G_ADD_PRIVATE for class private data
The old mechanism for creating class private data,
g_type_class_add_private(), is deprecated, and we should use
G_ADD_PRIVATE instead because it compiles down to faster code.

Removes some inconsistencies in naming, so that the ChimaraGlk*
parameter in methods that use the private data is now always named
"self".

While making this change, we rearrange some other code that peeked into
ChimaraGlk's private data, so that instead it calls private methods of
ChimaraGlk.

For safety, new code uses g_autofree where convenient, so we bump the
GLib requirement to 2.44.