Support CSS length with "pt" in CSS parser
Previously CSS like "font-size: 11pt" would cause a CSS parsing error.
Unfortunately GScanner only supports ending a numeric literal token on a
whitespace, so in order to parse something like "11pt" we have to
temporarily change the parser configuration to treat it like an identifier
and then restore the old definition of identifier when done.