Fix buffer underflow with finishing line input when none has been given
When finishing line input, we retrieve the entered text and chomp the
newline off the end. However, if the entered text is empty (which happens
when cancelling line input, for example) this would cause a buffer
underflow because we'd write one byte before the beginning of the string.
Fix this by only chomping the newline if the string's length is at least
1.