Fix use-after-free in UI messages awaiting a response
Previously the UI message would be freed at the end of ui_message_perform
in the UI thread, but then the response would still be read out of it in
the Glk thread. This would usually happen to work, but would occasionally
cause crashes or hangs.

Instead, set a flag in the UI message when it is being awaited so that we
don't free it until the response has been received in the Glk thread.