Fix emitting signals when glkunix_startup_code fails
We would not run glk_main() but exit the Glk thread immediately if
glkunix_startup_code() returns FALSE, signifying failure. However, in this
situation, neither the ::started nor ::stopped signals would be emitted.
We need to emit at least ::stopped, because programs may be relying on it
to know when to call gtk_main_quit(). And if we emit ::stopped, we should
emit ::started before it.