Sync new Glulxercise tests with glk-dev
Original location:
https://github.com/erkyrath/glk-dev/tree/master/unittests

Remove the Z5 tests which are mostly duplicates. (We don't need to test
the interpreters' implementations, we want to test the Glk library using
the Glulxe reference implementation.)

Also add the Regtest scripts that go along with each test, in order to run
these as part of automation. (Don't include any tests that don't have a
Regtest script.)