Retain startup args during Glk program lifetime
Previously we would free the startup argument strings right after
executing the Unix startup function in the Glk program. However, the Glk
program may hold on to these strings; Glulxe does, for one. Move them to
the ChimaraGlkPrivate data structure to ensure they last for the lifetime
of the Glk program.