Fix memory leaks when looking up cached image
In glk_image_get_info(), glk_image_draw(), and glk_image_draw_scaled(),
we'd check if an image of the correct size was already cached, but would
not free the lookup specifier.

We can allocate the lookup specifier on the stack anyway, instead of the
heap, so we don't need to worry about freeing it.