Fix memory leak reading byte file stream in glk_get_line_stream_uni()
If reading a stream of byte data from a file into a Unicode buffer, we
first need to read the bytes into a temporary byte buffer, then copy each
element into a 32-bit element in the buffer we return to the Glk program.
Previously, this temporary buffer would not be freed.