Fix race condition with releasing unicode line input buffers
It happened that a window could have handled the line input event on the
UI thread before glk_select() was called, and so the type of its pending
input request would be NONE by the time glk_select() needed to figure out
whether to release the buffer as a char buffer or unicode buffer. This
could cause the dispatch unregister callback to be called with the wrong
buffer.

So, put a new flag in the window struct, to indicate whether the last
line input request was unicode or not. This is a hacky solution; it would
be better if the event struct itself could indicate the type of the line
input request.