Skip unloading plugin if running address sanitizer
This is an unfortunate hack that _happens_ to make Glulxe work correctly
when restarted without unloading the glulxe.so plugin. (The problem is in
init_dispatch() which is not run a second time if it's already allocated
the static data once, but that's where the object registry is set up.)

This is needed in order to run things under the address sanitizer, because
otherwise it will identify Glulxe's static data as "leaks", which normally
would match the lines in the suppressions file, but without a stack trace
(because the module is unloaded) they cannot match.