Automate Glulxercise test suite
This runs the entire set of Zarf's Glulxercise tests, along with their
corresponding Regtest scripts, during 'meson test'. We add a runner
program which takes care of instantiating a ChimaraIF widget in a GTK
offscreen window, parsing the Regtest script, and performing the requested
checks.

Some of the checks are not currently possible (like checking graphics
output, or synthesizing hyperlink input.)