Add memory leak suppressions file
Both GLib and Glulxe allocate some static data. These allocations are not
memory leaks, because they remain in memory throughout the lifetime of the
process.

The suppression in prepare_glk_args() is unfortunate because that function
uses a scratch buffer which it frees only as needed when overwriting it
with a new allocation. However, we can't get any more granular with the
suppression rules, so there may be other memory leaks in that function
which aren't caught.

perform_saveundo() leaks on the Ubuntu image we use in CI for some reason,
but I can't reproduce it on my system.