Add CI address sanitizer job
With all of these fixes, it should be possible to run address sanitizer in
the automated tests, with the goal of catching memory leaks in every pull
request.

This requires some workarounds in Docker because it seems that the
available debuginfo package repo for GitHub Actions' Ubuntu 22.04 image
doesn't match exactly the versions available in the main package repo.