Implement graphics window character input
Character input in graphics windows is new in version 0.7.5 of the Glk
specification. It's relatively easy to enable; add key presses to the
events accepted by the graphics window's GtkDrawingArea, and allow the
graphics window to have the input focus.

Add a test, for which we also have to give glkunit-runner the ability to
send character input to the Glk program.