Hide windows when they have zero height or width
New in the 0.7.5 Glk spec is a note about zero-size windows. They should
disappear completely, including any borders that they have. This change
makes the size allocation code ignore zero-size windows, and in the case
of a pair window that is going to allocate zero size to one of its
children, it should instead give the entire space to its other child,
eliminating the border.

This is an interesting problem to test, but it can be done by testing
against a reference image. This mechanism might be useful for writing more
tests in the future, so we add a reftest runner (and image diffing code
copied from GTK's reftest suite.)