frotz: bool is a reserved word in C23
Starting from C23, bool is a reserved word, and some recent compilers
warn about this, causing builds to fail when compiling with unspecified
C standard (i.e. without an -std=cXX command-line argument) because
frotz defines its own bool type as an alias for int.

A possible solution would be to force compilation under an older
revision of the standard, but since bool has been a part of the standard
since C99, just use the standard header stdbool.h when building under
C99 or higher and leave the previous hand-rolled bool as a fallback for
ancient compilers.

This also requires the signature of the os_read_key in the header to use
the same type as the implementation (in glkmisc.c) for the second
argument.